requirements.txt
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.2
django-allauth==0.39.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
idna==2.8
oauthlib==3.0.1
pep8==1.7.1
Pillow==6.0.0
pycodestyle==2.5.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.21.0
requests-oauthlib==1.2.0
sqlparse==0.2.4
stripe==2.27.0
urllib3==1.24.2

boto3==1.10.49
dj-database-url==0.5.0
django-filter==2.2.0
django-heroku==0.3.1
django-model-utils==3.2.0
django-mssql==1.8
django-payu==0.5
django-phone-field==1.8.0
django-phonenumber-field==4.0.0
django-storages==1.8
django-stubs==1.4.0
djangorestframework==3.11.0
djangorestframework-stubs==1.1.0
gunicorn==20.0.4
Jinja2==2.10.3
s3transfer==0.2.1
requests==2.21.0
django-autocomplete-light
django-google-analytics-app
whitenoise
six==1.13.0
django-js-asset==1.0.0
django-widget-tweaks

django-import-export